requirements.txt
kafka-python
requests
streamlit
plotly